chore(chat): remover código comentado e componente não usado
