+++ frontend/src/features/services/components/Chat/Chat.tsx
@@ -3,15 +3,15 @@ import Input from "@/shared/components/ui/Inputs/Input";
 import ProfilePhoto from "@/shared/components/ui/ProfilePhoto";
 import clsx from "clsx";
 import { MdOutlineSend } from "react-icons/md";
-import type { ServicoDTO } from "../dtos/ServicoDTO";
+import type { ServicoDTO } from "../../dtos/ServicoDTO";
 import { toLocaleString } from "@/shared/utils/date.utils";
 import Title from "@/shared/components/ui/titles/Title";
 import { useAuth } from "@/features/auth/hooks/useAuth";
 import { useEffect, useRef } from "react";
 import { BackButton } from "@/shared/components/ui/buttons/BackButton";
 
-import { formatInTimeZone } from "date-fns-tz";
-import { ptBR } from "date-fns/locale";
+import CaixaMensagem from "./CaixaMensagem";
+import LimitadorLargura from "./LimitadorLargura";
 
 type ChatProps = {
   mensagem: string;
@@ -57,7 +57,7 @@ export const Chat = ({
   return (
     <div className="flex flex-col flex-1 min-h-0 w-full">
       <div className={clsx("p-4 border-y border-y-gray-300")}>
-        <WidthLimiter>
+        <LimitadorLargura>
           <section className="flex justify-between w-full">
             <Title className="text-start">
               <BackButton>{servico.nome}</BackButton>
@@ -74,11 +74,11 @@ export const Chat = ({
             />
             <p>{destinatario.nome}</p>
           </section>
-        </WidthLimiter>
+        </LimitadorLargura>
       </div>
 
       <section className={clsx("flex-1 overflow-y-auto p-4 bg-gray-100")}>
-        <WidthLimiter>
+        <LimitadorLargura>
           {mensagens.map((m, index) => (
             <CaixaMensagem
               key={index}
@@ -88,13 +88,13 @@ export const Chat = ({
             />
           ))}
           <div ref={messagesEndRef} />
-        </WidthLimiter>
+        </LimitadorLargura>
       </section>
 
       <section
         className={clsx("border-t border-t-gray-300 bg-white p-4 flex gap-4")}
       >
-        <WidthLimiter direction="row">
+        <LimitadorLargura direction="row">
           <Input
             placeholder="Digite sua mensagem"
             value={mensagem}
@@ -102,70 +102,8 @@ export const Chat = ({
             onKeyDown={handleKeyDown}
           />
           <Button icon={<MdOutlineSend />} onClick={enviarMensagem} />
-        </WidthLimiter>
+        </LimitadorLargura>
       </section>
-    </div>
-  );
-};
-
-const WidthLimiter = ({
-  children,
-  direction,
-}: {
-  children: React.ReactNode;
-  direction?: "row" | "col";
-}) => {
-  return (
-    <div
-      className={clsx(
-        "flex gap-4 px-4",
-        "md:px-10 max-w-280 mx-auto w-full",
-        direction === "row" ? "flex-row" : "flex-col",
-      )}
-    >
-      {children}
-    </div>
-  );
-};
-
-const CaixaMensagem = ({
-  mensagem,
-  isRemetente,
-  time,
-}: {
-  mensagem: string;
-  isRemetente: boolean;
-  time: string;
-}) => {
-  const timeZone = "America/Sao_Paulo";
-
-  return (
-    <div
-      className={clsx(
-        "flex w-full",
-        isRemetente ? "justify-end" : "justify-start",
-      )}
-    >
-      <div
-        className={clsx(
-          "flex flex-col rounded-xl p-2 max-w-[70%] break-all",
-          isRemetente ? "bg-secondary-100" : "bg-gray-300",
-        )}
-      >
-        <p className={clsx(isRemetente ? "self-end" : "self-start")}>
-          {mensagem}
-        </p>
-        <p
-          className={clsx(
-            "text-neutral-80",
-            isRemetente ? "self-end" : "self-start",
-          )}
-        >
-          {formatInTimeZone(time, timeZone, "HH:mm", {
-            locale: ptBR,
-          })}
-        </p>
-      </div>
     </div>
   );
 };
